third_party/WebKit/LayoutTests/fast/events/touch/gesture/touch-gesture-noscroll-body-expected.txt

This tests that a page cannot be scrolled with touch if its body has style overflow:hidden.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
first gesture
PASS document.documentElement.scrollTop is 0
PASS document.documentElement.scrollLeft is 0
PASS wheelEventsOccurred is 0
second gesture
PASS document.documentElement.scrollTop is 0
PASS document.documentElement.scrollLeft is 0
PASS wheelEventsOccurred is 0
PASS scrollEventsOccurred is 0
PASS successfullyParsed is true

TEST COMPLETE
